DHUD-193 Fixed some opacity issues (opacity is from 0 to 1 now, instead of 0 to 100), fixed error in the HUD (it stopped working if mouse was over the stat)

+++ DriveHUD.Application/ViewModels/Hud/HudStatSettingsViewModelInfo.cs
@@ -24,7 +24,8 @@ namespace DriveHUD.Application.ViewModels.Hud
         public StatInfo SelectedStatInfo { get; set; }
 
         public IEnumerable<StatInfo> SelectedStatInfoCollection { get; set; }
-        public int HudOpacity { get; set; }
+
+        public double HudOpacity { get; set; }
 
         public Action Save { get; set; }
